+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
 
+**v2.10.0**:
+- added update functionality to Environments, Workspace Spark Settings, Shortcuts, Job Schedules, ...
+- fixed issue where Workspaces did not load correctly if Folders were used
+
 **v2.9.5**:
 - improved logging when creating new treeview-items ([#33](/../../issues/33))
-
 
 **v2.9.4**:
 - fixed issue with special characters in workspace and/or item names ([#33](/../../issues/33))